helper file created, function moved

in progress - to be squashed


in progess
